coloring of top nav clicked links and some refactoring/splitting of header component
