+++ src/lib/layout/HeaderBottomNav.spec.js
@@ -1,0 +1,26 @@
+import HeaderBottomNav from './HeaderBottomNav.svelte';
+import {render, screen} from '@testing-library/svelte';
+
+describe('Nav should have dropdown sub menus', () => {
+  const navItems = [
+    {
+      key: 'navbar.about',
+      category: 'about',
+      children: ['navbar.about.team', 'navbar.about.values'],
+    },
+    {
+      key: 'navbar.using_data',
+      category: 'using_data',
+      children: ['navbar.using_data.projects', 'navbar.using_data.consulting'],
+    },
+  ];
+
+  test('toogles on message event', () => {
+    render(HeaderBottomNav, {props: {bot_nav: navItems}});
+
+    // Relies on a lot of configuration wrt. translations, meaning
+    // Language specific lookups. It would be nice if testing was more explict
+    expect(screen.getByText('Über uns')).toBeInTheDocument();
+    expect(screen.getByText('Daten nutzen')).toBeInTheDocument();
+  });
+});
